Share common convention plugin utilities in Ext.kt
