--- build-logic/src/main/kotlin/ProjectExt.kt
@@ -1,7 +1,0 @@
-import org.gradle.accessors.dm.LibrariesForLibs
-import org.gradle.api.Project
-import org.gradle.kotlin.dsl.findByType
-import org.gradle.kotlin.dsl.getByType
-
-val Project.libs: LibrariesForLibs
-  get() = extensions.findByType() ?: rootProject.extensions.getByType()
