+++ build-logic/src/main/kotlin/spotless.error-prone.gradle.kts
@@ -4,11 +4,6 @@ plugins {
   id("net.ltgt.errorprone")
 }
 
-// Groovy's String.toBoolean() -- what this used before the Kotlin DSL migration -- accepts "true",
-// "y" and "1", ignoring case. Kotlin's toBoolean() only accepts "true", so the short forms would
-// silently become no-ops.
-fun String.toBooleanGroovy(): Boolean = trim().lowercase() in setOf("true", "y", "1")
-
 tasks.withType<JavaCompile>().configureEach {
   options.errorprone {
     enabled = System.getenv("error-prone")?.toBooleanGroovy() == true
