+++ lib/build.gradle.kts
@@ -1,4 +1,3 @@
-import com.diffplug.spotless.changelog.gradle.ChangelogExtension
 import com.github.spotbugs.snom.Confidence
 
 plugins {
@@ -11,7 +10,7 @@ plugins {
 
 extra["artifactId"] = property("artifactIdLib")
 
-version = rootProject.the<ChangelogExtension>().versionNext
+version = spotlessChangelog.versionNext
 
 val needsGlue =
     listOf(
